Refactor user routes as blueprint
