+++ user/__init__.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+
+bp = Blueprint("user", __name__)
+from user import routes
